+++ src/app/(after-login)/addepigram/page.tsx
@@ -1,5 +1,3 @@
-import AddEpigramForm from '@/components/form/AddEpigramForm';
-
 export default function AddEpigramPage() {
-  return <AddEpigramForm />;
+  return <div>에피그램 생성 페이지입니다.</div>;
 }
